Refactor component folder structure and imports
Moved components from 'Manus' subfolders to more descriptive locations such as 'Headers', 'Sections', 'Views', and 'Icons'. Updated all relevant import paths in affected files to reflect the new structure for improved organization and maintainability.
